Bind MCPServer to all available interfaces
Originally we were binding to a single address on a single interface which,
thanks to LiSP's `get_lan_ip()` function, was the "default" address of the
underlying Operating System.

This is not so good on a system with more than one interface (e.g. WiFi and
Ethernet), and our Radio Receivers are on a network accessible through one of
the to a non-default interfaces.

Fortunately, python's `socket` module allows us to bind to a specified port on
*every* available/active interface, instead of having to bind to a single
interface.